// components/src/store/flow-definition.d.ts
export type ActionType =
  | 'execute_actions'
  | 'add_contact_urn'
  | 'add_contact_groups'
  | 'add_input_labels'
  | 'remove_contact_groups'
  | 'set_contact_channel'
  | 'set_contact_field'
  | 'set_contact_name'
  | 'set_contact_language'
  | 'set_contact_status'
  | 'set_run_result'
  | 'call_classifier'
  | 'call_resthook'
  | 'call_webhook'
  | 'call_llm'
  | 'open_ticket'
  | 'send_msg'
  | 'send_email'
  | 'send_broadcast'
  | 'enter_flow'
  | 'terminal'
  | 'start_session'
  | 'transfer_airtime'
  | 'split_by_airtime'
  | 'split_by_expression'
  | 'split_by_contact_field'
  | 'split_by_run_result'
  | 'split_by_run_result_delimited'
  | 'split_by_groups'
  | 'split_by_intent'
  | 'split_by_random'
  | 'split_by_resthook'
  | 'split_by_ticket'
  | 'split_by_scheme'
  | 'split_by_subflow'
  | 'split_by_webhook'
  | 'split_by_llm'
  | 'split_by_llm_categorize'
  | 'wait_for_response'
  | 'wait_for_menu'
  | 'wait_for_dial'
  | 'wait_for_digits'
  | 'wait_for_audio'
  | 'wait_for_video'
  | 'wait_for_location'
  | 'wait_for_image'
  | 'request_optin'
  | 'missing'
  | 'say_msg'
  | 'play_audio';

export interface Action {
  type: ActionType;
  uuid: string;
}
// ...
export interface Exit {
  uuid: string;
  destination_uuid?: string;
}

export type Hint = {
  type: 'digits' | 'audio' | 'image' | 'video' | 'location';
  count?: number;
};

export interface Timeout {
  category_uuid: string;
  seconds: number;
}

export interface Wait {
  type: 'msg' | 'dial';
  timeout?: Timeout;
  hint?: Hint;
  phone?: string;
  dial_limit_seconds?: number;
  call_limit_seconds?: number;
}

export interface Category {
  uuid: string;
  name: string;
  exit_uuid: string;
}

export interface Case {
  uuid: string;
  type:
    | 'has_any_word'
    | 'has_number_between'
    | 'has_group'
    | 'has_only_phrase'
    | 'has_only_text'
    | 'has_text'
    | string;
  arguments: string[];
  category_uuid: string;
}

export interface Router {
  type: 'switch' | 'random';
  result_name?: string;
  operand?: string;
  default_category_uuid?: string;
  cases?: Case[];
  categories: Category[];
  wait?: Wait;
}

export interface Node {
  uuid: string;
  actions: Action[];
  exits: Exit[];
  router?: Router;
}